+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.7.1...HEAD)
 
+- [JSHint] Show runtime versions [#407](https://github.com/sider/runners/pull/407)
+
 ## 0.7.1
 
 [Full diff](https://github.com/sider/runners/compare/0.7.0...0.7.1)